pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "solar-miner"
version = "0.1.0"
description = "Throttle Antminer S19s to consume only surplus rooftop solar"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "sqlite-utils>=3.36",
]

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]
